Split Windows hsMessageBox into its own file
